# Backend/api/docs/api.yaml
openapi: 3.0.3
info:
  title: My API
  version: 1.0.0
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
    AppClientAuth:
      type: apiKey
      in: header
      name: X-App-Secret
  schemas:
...
    def-1:
      additionalProperties: false
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
      type: object
      title: errorResponse
...
    def-3:
      description: OK
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
      title: SuccessMessageResponse
...
    def-6:
      type: object
      required:
        - id
        - username
        - avatar
      properties:
        id:
          type: integer
        username:
          type: string
        avatar:
          type:
            - "null"
            - string
        is_active:
          type: boolean
        oauth:
          type: boolean
      title: publicUserObject
...
paths:
...
  /api/users/{id}:
    get:
      tags:
        - user
      description: Returns an object of an user using the id passed in parameters.
        `This endpoint requires client authentication.`
      parameters:
        - schema:
            type: integer
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-6"
        "400":
          description: "Bad Request: Invalid input or missing fields"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-1"
        "401":
          description: "Unauthorized: Invalid credentials"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-1"
        "404":
          description: "Not Found: User not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-1"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                description: Internal Server Error
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-1"
...
  /api/users/me/avatar:
    put:
      tags:
        - user
      description: Modifies the avatar of the user. `This endpoint requires client AND
        user authentication.`
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - avatar
              properties:
                avatar:
                  type: string
                  format: binary
        required: true
      security: *a1
      responses:
        "200":
          description: Updated avatar successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-3"
        "400":
          description: "Bad Request: Invalid input or missing fields"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-1"
        "401":
          description: "Unauthorized: Invalid credentials"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-1"
        "413":
          description: Payload too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-1"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-1"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-1"
    delete:
      tags:
        - user
      description: Deletes the avatar of the user. `This endpoint requires client AND
        user authentication.`
      security: *a1
      responses:
        "204":
          description: "Success: Avatar deleted successfully"
        "401":
          description: "Unauthorized: Invalid credentials"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-1"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-1"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-1"
...
  /api/users/auth/reset-password:
    post:
      tags:
        - auth
      description: Sends a password reset email to the user's email address. If the
        email is not found a 200 is still returned
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
        required: true
      responses:
        "200":
          description: "Success: Password reset email sent"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-3"
        "400":
          description: "Bad Request: Invalid input or missing fields"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-1"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-1"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-1"
    get:
      tags:
        - auth
      description: "Verifies a user's password reset token and redirects to the reset
        password page where they can reset their password "
      parameters:
        - schema:
            type: string
          in: query
          name: token
      responses:
        "302":
          description: Redirects to the reset password page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-3"
        "400":
          description: "Bad Request: Invalid input or missing fields"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-1"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-1"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-1"
...